requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-language-field==0.0.3
django-multiselectfield==0.1.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
